sqlite db setup

+++ server/models/Movie.py
@@ -7,5 +7,5 @@ from sqlmodel import Field, Session, SQLModel, create_engine, select
 class Movie(SQLModel, table=True):
     id: int | None = Field(default=None, primary_key=True)
     name: str = Field(index=True)
-    rating: int = Field(default=None, index=True)
+    rating: float = Field(default=None, index=True)  # Changed from int to float to match schema
     plot: str
